Move MockSocket into Mock
